+++ src/api/controllers/claim-v1-router.js
@@ -77,7 +77,7 @@ class ClaimController {
       return;
     }
     ClaimService
-      .createWithClaimEntry(req.body.jwtEncoded, res.locals.tokenIssuer)
+      .createWithClaimEntry(req.body.jwtEncoded, res.locals.tokenPayload, res.locals.tokenIssuer)
       // no need to check for visible data because they sent it
       .then(r => {
         const result = r.claimId